zbd: return eof if file closing
If the target device file is marked as closing, zone locks taken
within zbd_adjust_block() will not be released as the IOs will not be
issued. This can potentially trigger a deadlock when a run completes as
some jobs may forever wait for a zone lock to be released.
Avoid this by returning eof in zbd_adjust_block() if the target file is
marked as closing.

Signed-off-by: Damien Le Moal <damien.lemoal@wdc.com>